Fixed field type and batch generator.

+++ src/main/java/uk/ac/ebi/spot/ontotools/curation/domain/ImportLogBatch.java
@@ -29,7 +29,7 @@ public class ImportLogBatch {
 
     private DateTime timestamp;
 
-    private long totalTime;
+    private int totalTime;
 
     private int totalCount;
